requirements.txt
bitstring==3.1.9
chardet==4.0.0
jsonpath==0.82
jsonpath_rw==1.4.0
tabulate==0.8.9
websocket_client==1.3.2
PyPDF2
python-docx
pandas
Pillow
pdf2image
python-pptx
jieba
pdfplumber
openpyxl
xlrd
natsort
moviepy
Flask
beautifulsoup4